issue #138 Deduplicate some code
